Refactor AI with better FSM and FSM builder (#41)
This PR separates AI Final State Machine into its own class (alongside with Timers) and adds a builder to create new FSM states. Because I don't know how to make it templated, the FSM is tightly coupled to the list of states defined in AiBrain.

**How the FSM works?**

At the top level, each state of the FSM has a list of conditions, that are always evaluated in order. If any of these conditions returns true, FSM transitions to state associated with the condition.

After evaluating conditions, if all of them were false, the FSM executes the default behaviour of the state and if there is a default transition, it jumps to the next state.

**How the builder works**

If you import If, Do and DoNothing symbols from fsm-builder module, you can easily build new states:

```
 If(someCondition).goTo(stateCode)
.orIf(anotherCondition).goTo(anotherState)
.otherwiseDo(defaultLogic).andLoop()
```

```
 If(someCondition).goTo(stateCode)
.otherwiseDo(defaultLogic).thenGoTo(anotherState)
```

```
Do(someLogic).thenGoTo(someState)
```

Note that if you try to create a looping state without any conditions, it will fire a runtime assertion (which is covered implicitly in unit tests).

The problem is, that function passed to the builder are not function calls, they are function references. Which means that `this` will inevitably get screwed all over. This is why all the methods that can be used by the FSM need to accept parameter of `AiBrain` which is used instead of `this`. Also the second mandatory parameter is `GameContext`.

Methods, that are called from within these top level ones can use `this` normally (they are marked by NON FMS METHODS comment in the code). If you need extra parameters to your methods, or want to pass a parameter directly to the FSM builder, you need to write you own curry function. `and`, `not`, `doNothing` and `timerEnded` are examples of such curries.