Feature: Projectiles get faster and bigger as the game goes on (#74)
Name of the PR is self descriptive. Game now has concept of ramping up difficulty. GameContext now tracks total elapsed time in each game that can be used to current difficulty factor from GameMath.

There are new settings for when the difficulty starts ramping up (currently after 15s) and when it stops (is stuck at max, currently 60s). From that, scaling for projectile size and speed is computed.

Renderer had to be updated so it now asks GameObjects for their scale and can scale sprites accordingly.