Fix AI not working at 30FPS (#81)
The main issue was that AI doesn't update with delta time but rather
with FIXED_FRAME_TIME (we could actually remove delta time completely
since everything is fixed anyways).

Also there is a problem that certain hand tuned constants were not
adjusted at 30FPS. The angle difference threshold was the biggest
offender as it prevented AI to reach any target angle and thus it was
stuct in one state.

There are few more variables that would need some hand tuning and 30FPS
is just too big of a step for certain things to work properly, but at
least it behaves now sort of as expected.

I also swapped buttons for Start game and Back to menu. It made more
sense to have Start game on the right.